Moved EB-specific classes to Eventbrite class (formerly EventBriteAPI) Renamed EB-specific classes to EB* Altered EventsFeedMetaData to take a name parameter as well as a token; this can then be used for multiple EB feeds Actioned other numerous points from team PR comments (created EventHandler to handle identification of "live" events)
